Move MainActivity.java to project root
