+++ MainActivity.java
@@ -1,10 +1,12 @@
 package com.aself.student.librarycard;
 
+import android.app.Activity;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.os.Handler;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.os.AsyncTask;
 import android.util.Base64;
 import android.view.View;
 import android.widget.Button;
@@ -22,17 +24,10 @@ import java.net.InetAddress;
 import java.net.Socket;
 
 public class MainActivity extends AppCompatActivity {
-    static final String SERVER_IP = "192.168.1.115";
-    static final int SERVER_PORT = 4303;
-
-    Handler handler = new Handler();
-    static Socket socket;
-    PrintWriter printWriter;
     ImageView image;
     EditText inputbox;
     Button button;
     TextView errorbox;
-
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -49,97 +44,56 @@ public class MainActivity extends AppCompatActivity {
     @Override
     protected void onStart() {
         super.onStart();
-        Thread clientThread = new Thread(new ClientThread());
-        errorbox.setText("After thread instantiated");
-        clientThread.start();
     }
 
     @Override
     protected void onStop() {
         super.onStop();
-        try {
-            socket.shutdownInput();
-            socket.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
     }
 
     public void onClick(View view) {
-        handler.post(new Runnable() {
-            @Override
-            public void run() {
-                printWriter.println(inputbox.getText().toString());
-            }
-        });
+		new TalkToServerTask().execute(inputbox.getText().toString());
     }
 
-    public class ClientThread implements Runnable {
-        public void run() {
-            errorbox.setText("ClientThread.run() called");
-            try {
-                errorbox.setText("before the getByName call");
-                InetAddress serverAddr = InetAddress.getByName(SERVER_IP);
-                errorbox.setText("after getByName; before socket");
-                socket = new Socket(serverAddr,SERVER_PORT);
-                errorbox.setText("before the try statement");
-                try {
-                    errorbox.setText("printWriter about to be defined");
-                    printWriter = new PrintWriter(new BufferedWriter(new OutputStreamWriter(socket.getOutputStream())),true);
-                    errorbox.setText("PrintWriter Defined");
-                    BufferedReader br = new BufferedReader(new InputStreamReader(socket.getInputStream()));
-                    try {
-                        String line = null;
-                        while ((line = br.readLine()) != null) {
-                            final String strReceived = line;
-                            handler.post(new Runnable() {
-                                @Override
-                                public void run() {
-                                    byte[] imageBytes = Base64.decode(strReceived, Base64.DEFAULT);
-                                    Bitmap decodedImage = BitmapFactory.decodeByteArray(imageBytes, 0, imageBytes.length);
-                                    image.setImageBitmap(decodedImage);
-                                }
-                            });
-                        }
-                        handler.post(new Runnable() {
-                            @Override
-                            public void run() {
-                                errorbox.setText("Client Disconnected");
-                            }
-                        });
-                    } catch (Exception e) {
-                        final String error = e.getLocalizedMessage();
-                        handler.post(new Runnable() {
-                            @Override
-                            public void run() {
-                                errorbox.setText(error);
-                            }
-                        });
-                    }
-                } catch (Exception e) {
-                    final String error = e.getLocalizedMessage();
-                    handler.post(new Runnable() {
-                        @Override
-                        public void run() {
-                            errorbox.setText(error);
-                        }
-                    });
-                }
-                handler.post(new Runnable() {
-                    @Override
-                    public void run() {
-                        errorbox.setText("Connection Closed");
-                    }
-                });
-            } catch (Exception e) {
-                final String error = e.getLocalizedMessage();
-                handler.post(new Runnable() {
-                    @Override
-                    public void run() {
-                        errorbox.setText(error);
-                    }
-                });
-            }
-        }
+    public class TalkToServerTask extends AsyncTask<String, Void, Bitmap> {
+		private static final String SERVER_IP = "192.168.1.115";
+		private static final int SERVER_PORT = 4303;
+
+		// This method is called by execute() on a new Thread()
+		// Don't mess with the UI here or Android will be angry
+        @Override
+        public Bitmap doInBackground(String... params) {
+			// FIXME you need to actually do whatever with your errors
+			InetAddress serverAddr;
+			Bitmap decodedImage = null;
+			try {
+                serverAddr = InetAddress.getByName(SERVER_IP);
+				Socket socket = new Socket(serverAddr,SERVER_PORT);
+				PrintWriter printWriter = new PrintWriter(new BufferedWriter(new OutputStreamWriter(socket.getOutputStream())),true);
+				BufferedReader br = new BufferedReader(new InputStreamReader(socket.getInputStream()));
+				// Send librarycard
+				String librarycard = params[0];
+				printWriter.println(librarycard);
+				// FIXME I wasn't super sure what exactly was going on here, but you
+				// need to return decodedImage when it is all good
+				String line = null;
+				while ((line = br.readLine()) != null) {
+					final String strReceived = line;
+					byte[] imageBytes = Base64.decode(strReceived, Base64.DEFAULT);
+					decodedImage = BitmapFactory.decodeByteArray(imageBytes, 0, imageBytes.length);
+				}
+			} catch (Exception e) {
+                // FIXME handle error however you want
+			}
+			// return decodedImage here
+			return decodedImage;
+		}
+
+		// This method is called on the main/UI Thread
+		// Do whatever UI interaction you need here
+		@Override
+		public void onPostExecute(Bitmap decodedImage){
+			image.setImageBitmap(decodedImage);
+		}
     }
 }
